delete

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -14,6 +14,8 @@ if __name__ == "__main__":
                              av[1], av[2], av[3]), pool_pre_ping=True)
     Session = sessionmaker(bind=engine)
     session = Session()
-    for ins in session.query(State).filter(State.name.like('%a%')):
-        session.delete(ins)
-        session.commit()
+    instances = session.query(State)
+    for instance in instances:
+        if 'a' in instance.name:
+            session.delete(ins)
+            session.commit()
